Add option for disabling block magnetism during manipulation.
This mod is enabled by setting the attribute 'disable_block_magnets_during_manipulation' on the virtual magnetism plugin to true. In this case, the magnet forces between blocks will not be applied while a BuilderBot end effector is interacting with them. The mod is useful for allowing blocks to be removed from a structure.