Remove the environment XML node and update the Pi-Puck and drone to take command line arguments.
This commit removes the parsing of the environment node and instead relies on commmand line configuration. For example, the drone can be started with the command:
argos3 --config drone.argos --id drone100 --pixhawk /dev/ttyS1:921600 --router 192.168.1.2:4950

The Pi-Puck is the same except it ignores the Pixhawk argument. Arguments can also be provide with a single dash followed by the first letter of the parameter (e.g. -i drone100).