Remove LED detection from the drone camera sensor.
LED detection has been broken in hardware since switching to a JPEG capture format. This commit completely removes this functionality so that the control interface and the simulation implementation are consistent with the hardware implementation. LED detection could be added back by using two `std::optional<::image_u8_t>` for the UV channels, which is initialized from the JPEG buffer the first time `DetectLED` is called.